Gentle Anchors 1.2.4
User reported issue with some links to other pages that used anchors
would not allow them to go to the page. The issue was with the
matching of the current page. When coming from say the home page,
then linking to a sub-page, the link would match because of allowing
for too many potential matches.